add connection request route
