+++ packages/desktop/views/dashboard/drawers/dapp-config/dapp-config-route.enum.ts
@@ -1,6 +1,7 @@
 export enum DappConfigRoute {
     ConnectedDapps = 'connectedDapps',
-    ConfirmConnection = 'ConfirmConnection',
-    DappDetails = 'DappDetails',
+    ConnectionRequest = 'connectionRequest',
+    ConfirmConnection = 'confirmConnection',
+    DappDetails = 'dappDetails',
     InputCode = 'inputCode',
 }
