+++ packages/desktop/views/dashboard/drawers/dapp-config/views/index.ts
@@ -1,4 +1,4 @@
-export { default as ConfirmConnectionDrawer } from './ConfirmConnectionDrawer.svelte'
+export { default as ConfirmConnectionDrawer } from './ConnectionRequestDrawer.svelte'
 export { default as ConnectedDappsDrawer } from './ConnectedDappsDrawer.svelte'
 export { default as DappDetailsDrawer } from './DappDetailsDrawer.svelte'
 export { default as InputConnectionCodeDrawer } from './InputConnectionCodeDrawer.svelte'
